Add address book storage management for desktop and extension clients
Introduce a new addressBook storage module to handle address book operations, including creating, updating, retrieving, and deleting address book items. Refactor existing code to utilize this new module, enhancing code organization and simplifying the core storage interface. Remove obsolete address book functions from the storage index files.
